backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "The Recommendation system"
authors = ["atabekdemurtaza <mychannel21war@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
drf-nested-routers = "^0.94.1"
faker = "^28.1.0"
pandas = "^2.2.2"
joblib = "^1.4.2"
numpy = "^1.21.0"
scikit-learn = "1.3.2"
nltk = "^3.9.1"
prettytable = "^3.11.0"
scipy = "^1.14.1"
imbalanced-learn = "^0.12.3"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"
django-cors-headers = "^4.4.0"
django-environ = "^0.11.2"
django-extensions = "^3.2.3"
pydotplus = "^2.0.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
